[refactor] some refactor
