+++ src/main.rs
@@ -14,10 +14,10 @@ fn main() ->Result<(),opencv::Error>{
     video_stream.moving_object_detection(3, 20, 30)?;
     let mut my_camera = stream::Stream::from_camera();
     // my_camera.moving_object_detection(6, 60, 30)?;
-    let _ =my_camera.face_detection();
-    let _ =my_camera.body_detection();
-    // my_camera.camera();
-    // my_camera.capture_frame();
+    my_camera.face_detection()?;
+    my_camera.body_detection()?;
+    // my_camera.camera()?;
+    // my_camera.capture_frame()?;
 
     Ok(())
 }
